adjusted sizing of main container & item container
The main problem with adjusting the sizing of the main container was the item div itself. The styling for it made it so it occupied a bunch of height so when you would adjust the width of the main-container the image would not resize properly due to the height being messed up by the item div.